+++ lib/sorting.dart
@@ -4,4 +4,5 @@
 /// A collection of list sorting algorithms.
 library algorithmic.sorting;
 
+export 'src/sorting/bubble.dart';
 export 'src/sorting/selection.dart';